Refactor tracing server
